Revert "refactor(bank-sync-tests): drop harness barrel, import from source files"
This reverts commit d42590ec724a56eefb720a30d65c4ef877f14b53.
