+++ tests/bank-sync-tests/src/scenarios/consolidation/transfer-pair-by-amount.test.ts
@@ -2,12 +2,7 @@ import { describe, expect, it } from 'vitest';
 
 import { TransactionConsolidationTypeEnum } from '@budgie/contracts';
 
-import { fetchCanonicalsOfType } from '../../harness/fetch-canonicals';
-import { fetchTransactionById } from '../../harness/fetch-by-id';
-import { findMccByCode } from '../../harness/find-mcc';
-import { seedBankExpense, seedBankIncome } from '../../harness/seed-bank-pair';
-import { seedAccountPair } from '../../harness/seed-transfer-pair-fixture';
-import { setupScenario } from '../../harness/setup-scenario';
+import { fetchCanonicalsOfType, fetchTransactionById, findMccByCode, seedAccountPair, seedBankExpense, seedBankIncome, setupScenario } from '../../harness';
 
 import { transferConsolidationService } from '@app/sync/service/transfer-consolidation.service';
