+++ tests/bank-sync-tests/src/scenarios/consolidation/unconsolidate-restores-sources.test.ts
@@ -3,11 +3,7 @@ import { eq } from 'drizzle-orm';
 
 import { TransactionConsolidationTypeEnum, TransactionEntityTable, TransactionEntryEntityTable } from '@budgie/contracts';
 
-import { fetchCanonicalsOfType } from '../../harness/fetch-canonicals';
-import { fetchTransactionById } from '../../harness/fetch-by-id';
-import { seedTransferPairFixture } from '../../harness/seed-transfer-pair-fixture';
-import { setupScenario } from '../../harness/setup-scenario';
-import { testDb } from '../../harness/setup';
+import { fetchCanonicalsOfType, fetchTransactionById, seedTransferPairFixture, setupScenario, testDb } from '../../harness';
 
 import { transactionService } from '@app/transaction/service/transaction.service';
 import { transferConsolidationService } from '@app/sync/service/transfer-consolidation.service';
