+++ tests/bank-sync-tests/src/scenarios/monobank/pagination-cursor-advances.test.ts
@@ -5,10 +5,7 @@ import { http, HttpResponse } from 'msw';
 import { BankSyncEntityTable, ExternalSourceEnum, TransactionEntityTable } from '@budgie/contracts';
 import type { MonobankTransactionApiInterface } from '@budgie/bank-sync';
 
-import { buildMonobankTx } from '../../harness/monobank-fixtures';
-import { setupMonobankFixture } from '../../harness/setup-monobank-fixture';
-import { setupScenario } from '../../harness/setup-scenario';
-import { testDb } from '../../harness/setup';
+import { buildMonobankTx, setupMonobankFixture, setupScenario, testDb } from '../../harness';
 import { monobankServer } from '../../harness/monobank-server';
 
 import { monobankSyncService } from '@app/sync/service/monobank-sync.service';
